requirements.txt
pandas>=1.5.0
google-genai>=0.2.0
anthropic>=0.18.0
requests>=2.28.0
Pillow>=9.0.0

